requirements.txt
simpleitk~=2.2.1
numpy~=1.24.3
natsort~=8.3.1
matplotlib~=3.7.1
torchvision~=0.15.0
pillow~=9.4.0
torchio~=0.18.91
monai~=1.1.0
pandas~=2.0.1
tensorboardx~=2.6
torchmetrics~=0.11.4
tqdm~=4.65.0
easydict~=1.10
imageio~=2.28.1
scipy>=1.8,<1.9.2
scikit-image~=0.20.0
deepspeed~=0.9.2
blobfile~=2.0.2
mpi4py~=3.1.4
opencv-python
easydict
blobfile
mpi4py
matplotlib